Refactor transaction construction across multiple modules to utilize `buildTransactionRequest` for improved consistency and maintainability. This change simplifies transaction request creation by centralizing gas estimation and options handling.
