+++ src/concentratedLiquidity/provision.ts
@@ -2,6 +2,7 @@ import { ethers, ContractReceipt } from "ethers";
 import { BatchLPDetails } from "./positionViewer";
 import { TransactionOptions } from "../types";
 import orderbookAbi from "../../abi/OrderBook.json";
+import buildTransactionRequest from "src/utils/txConfig";
 
 export abstract class PositionProvider {
     /**
@@ -100,45 +101,12 @@ export abstract class PositionProvider {
             false
         ]);
 
-        const tx: ethers.providers.TransactionRequest = {
-            to: contractAddress,
+        return buildTransactionRequest({
             from: address,
+            to: contractAddress,
+            signer,
             data,
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && { maxFeePerGas: txOptions.maxFeePerGas }),
-            ...(txOptions?.maxPriorityFeePerGas && { maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas
-                ? signer.provider!.getGasPrice()
-                : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(
-                    txOptions.priorityFee.toString(),
-                    "gwei"
-                );
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            txOptions
+        });
     }
 }
