+++ src/margin/deposit.ts
@@ -12,6 +12,7 @@ import { TransactionOptions } from "src/types";
 // ============ Config Imports ============
 import erc20Abi from "../../abi/IERC20.json";
 import marginAccountAbi from "../../abi/MarginAccount.json";
+import buildTransactionRequest from "src/utils/txConfig";
 
 export abstract class MarginDeposit {
     static async deposit(
@@ -85,54 +86,16 @@ export abstract class MarginDeposit {
             amount,
         ]);
 
-        const tx: ethers.providers.TransactionRequest = {
+        const value = tokenAddress === ethers.constants.AddressZero ? amount : BigNumber.from(0);
+
+        return buildTransactionRequest({
             to: marginAccountAddress,
             from: address,
             data,
-            value:
-                tokenAddress === ethers.constants.AddressZero
-                    ? amount
-                    : BigNumber.from(0),
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && {
-                maxFeePerGas: txOptions.maxFeePerGas,
-            }),
-            ...(txOptions?.maxPriorityFeePerGas && {
-                maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas,
-            }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas
-                ? signer.provider!.getGasPrice()
-                : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(
-                    txOptions.priorityFee.toString(),
-                    "gwei"
-                );
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            value,
+            txOptions,
+            signer
+        });
     }
 
     static async estimateGas(
