+++ src/market/ioc.ts
@@ -13,6 +13,7 @@ import { MarketParams, MARKET, TransactionOptions } from "../types";
 // ============ Config Imports ============
 import orderbookAbi from "../../abi/OrderBook.json";
 import erc20Abi from "../../abi/IERC20.json";
+import buildTransactionRequest from "src/utils/txConfig";
 
 export abstract class IOC {
     /**
@@ -153,55 +154,19 @@ export abstract class IOC {
             marketParams.quoteAssetDecimals
         );
 
-        const tx: ethers.providers.TransactionRequest = {
+        const value = !isMargin &&
+        marketParams.quoteAssetAddress === ethers.constants.AddressZero
+            ? parsedQuoteSize
+            : BigNumber.from(0);
+
+        return buildTransactionRequest({
             to: orderbookAddress,
             from: address,
             data,
-            value:
-                !isMargin &&
-                marketParams.quoteAssetAddress === ethers.constants.AddressZero
-                    ? parsedQuoteSize
-                    : BigNumber.from(0),
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && {
-                maxFeePerGas: txOptions.maxFeePerGas,
-            }),
-            ...(txOptions?.maxPriorityFeePerGas && {
-                maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas,
-            }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas
-                ? signer.provider!.getGasPrice()
-                : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(
-                    txOptions.priorityFee.toString(),
-                    "gwei"
-                );
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            value,
+            txOptions,
+            signer
+        });
     }
 
     /* @dev Constructs a market sell transaction.
@@ -241,56 +206,20 @@ export abstract class IOC {
             size,
             marketParams.baseAssetDecimals
         );
+        
+        const value = !isMargin &&
+        marketParams.baseAssetAddress === ethers.constants.AddressZero
+            ? parsedSize
+            : BigNumber.from(0);
 
-        const tx: ethers.providers.TransactionRequest = {
+        return buildTransactionRequest({
             to: orderbookAddress,
             from: address,
             data,
-            value:
-                !isMargin &&
-                marketParams.baseAssetAddress === ethers.constants.AddressZero
-                    ? parsedSize
-                    : BigNumber.from(0),
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && {
-                maxFeePerGas: txOptions.maxFeePerGas,
-            }),
-            ...(txOptions?.maxPriorityFeePerGas && {
-                maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas,
-            }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas
-                ? signer.provider!.getGasPrice()
-                : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(
-                    txOptions.priorityFee.toString(),
-                    "gwei"
-                );
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            value,
+            txOptions,
+            signer
+        });
     }
 }
