+++ src/utils/approve.ts
@@ -7,6 +7,7 @@ import { TransactionOptions } from "../types";
 // ============ Internal Imports ============
 import { extractErrorMessage } from "../utils";
 import erc20Abi from "../../abi/IERC20.json";
+import buildTransactionRequest from "./txConfig";
 
 const getOwnerAddress = async (
     providerOrSigner: ethers.providers.JsonRpcProvider | ethers.Signer
@@ -41,49 +42,13 @@ export async function constructApproveTransaction(
         size,
     ]);
 
-    const tx: ethers.providers.TransactionRequest = {
+    return buildTransactionRequest({
         to: tokenContractAddress,
         from: address,
         data,
-        ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-        ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-        ...(txOptions?.maxFeePerGas && {
-            maxFeePerGas: txOptions.maxFeePerGas,
-        }),
-        ...(txOptions?.maxPriorityFeePerGas && {
-            maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas,
-        }),
-    };
-
-    const [gasLimit, baseGasPrice] = await Promise.all([
-        !tx.gasLimit
-            ? signer.estimateGas({
-                  ...tx,
-                  gasPrice: ethers.utils.parseUnits("1", "gwei"),
-              })
-            : Promise.resolve(tx.gasLimit),
-        !tx.gasPrice && !tx.maxFeePerGas
-            ? signer.provider!.getGasPrice()
-            : Promise.resolve(undefined),
-    ]);
-
-    if (!tx.gasLimit) {
-        tx.gasLimit = gasLimit;
-    }
-
-    if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-        if (txOptions?.priorityFee) {
-            const priorityFeeWei = ethers.utils.parseUnits(
-                txOptions.priorityFee.toString(),
-                "gwei"
-            );
-            tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-        } else {
-            tx.gasPrice = baseGasPrice;
-        }
-    }
-
-    return tx;
+        txOptions,
+        signer
+    });
 }
 
 /**
