+++ src/vault/vault.ts
@@ -9,6 +9,7 @@ import { getTokenDecimals, approveToken } from "../utils";
 import vaultAbi from "../../abi/Vault.json";
 import marginAbi from "../../abi/MarginAccount.json";
 import erc20Abi from "../../abi/IERC20.json";
+import buildTransactionRequest from "src/utils/txConfig";
 
 export abstract class Vault {
     static async deposit(
@@ -331,42 +332,14 @@ export abstract class Vault {
                 ? amount2
                 : BigNumber.from(0);
 
-        const tx: ethers.providers.TransactionRequest = {
+        return buildTransactionRequest({
             to: vaultAddress,
             from: address,
             data,
             value: txValue,
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && { maxFeePerGas: txOptions.maxFeePerGas }),
-            ...(txOptions?.maxPriorityFeePerGas && { maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas ? signer.provider!.getGasPrice() : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(txOptions.priorityFee.toString(), "gwei");
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            txOptions,
+            signer
+        });
     }
 
     static async constructWithdrawTransaction(
@@ -381,40 +354,12 @@ export abstract class Vault {
 
         const data = vaultInterface.encodeFunctionData("withdraw", [shares, fromAddress, fromAddress]);
 
-        const tx: ethers.providers.TransactionRequest = {
+        return buildTransactionRequest({
             to: vaultAddress,
             from: fromAddress,
             data,
-            ...(txOptions?.nonce !== undefined && { nonce: txOptions.nonce }),
-            ...(txOptions?.gasLimit && { gasLimit: txOptions.gasLimit }),
-            ...(txOptions?.gasPrice && { gasPrice: txOptions.gasPrice }),
-            ...(txOptions?.maxFeePerGas && { maxFeePerGas: txOptions.maxFeePerGas }),
-            ...(txOptions?.maxPriorityFeePerGas && { maxPriorityFeePerGas: txOptions.maxPriorityFeePerGas }),
-        };
-
-        const [gasLimit, baseGasPrice] = await Promise.all([
-            !tx.gasLimit
-                ? signer.estimateGas({
-                      ...tx,
-                      gasPrice: ethers.utils.parseUnits("1", "gwei"),
-                  })
-                : Promise.resolve(tx.gasLimit),
-            !tx.gasPrice && !tx.maxFeePerGas ? signer.provider!.getGasPrice() : Promise.resolve(undefined),
-        ]);
-
-        if (!tx.gasLimit) {
-            tx.gasLimit = gasLimit;
-        }
-
-        if (!tx.gasPrice && !tx.maxFeePerGas && baseGasPrice) {
-            if (txOptions?.priorityFee) {
-                const priorityFeeWei = ethers.utils.parseUnits(txOptions.priorityFee.toString(), "gwei");
-                tx.gasPrice = baseGasPrice.add(priorityFeeWei);
-            } else {
-                tx.gasPrice = baseGasPrice;
-            }
-        }
-
-        return tx;
+            txOptions,
+            signer
+        });
     }
 }
